+++ src/main-code/conversion.ts
@@ -13,7 +13,7 @@ import { transformMutliLineText, transformText } from './transform';
  * @since 2024-04-04
  */
 export function caseConversion(targetCase: SupportCase, str: string, eol: EOL, cutText: Array<TransformTextResult> | undefined = undefined): string {
-    let spaceCharacter: '-' | '_' | ' ' | undefined = undefined;
+    let spaceCharacter: '-' | '_' | ' ' | '.' | undefined = undefined;
     switch (targetCase) {
         default:
         case SupportCase.CAMEL_CASE:        // 小驼峰(驼峰)命名
@@ -37,6 +37,12 @@ export function caseConversion(targetCase: SupportCase, str: string, eol: EOL, c
         case SupportCase.SPACE_PASCAL_CASE: // 空格分隔 + 大驼峰(帕斯卡)命名
         case SupportCase.SPACE_UPPER_CASE:  // 空格分隔 + 全大写命名
             spaceCharacter = ' ';
+            break;
+        case SupportCase.DOT_CASE:          // 点分隔命名
+        case SupportCase.DOT_CAMEL_CASE:    // 点分隔 + 小驼峰(驼峰)命名
+        case SupportCase.DOT_PASCAL_CASE:   // 点分隔 + 大驼峰(帕斯卡)命名
+        case SupportCase.DOT_UPPER_CASE:    // 点分隔 + 全大写命名
+            spaceCharacter = '.';
             break;
         case SupportCase.LOWER_CASE:        // 全小写
             return str.toLowerCase();
@@ -82,6 +88,7 @@ export function caseConversion(targetCase: SupportCase, str: string, eol: EOL, c
                 case SupportCase.SNAKE_CAMEL_CASE:  // 下划线(蛇形) + 小驼峰(驼峰)命名
                 case SupportCase.KEBAB_CAMEL_CASE:  // 中划线(连字符/脊柱式) + 小驼峰(驼峰)命名
                 case SupportCase.SPACE_CAMEL_CASE:  // 空格分隔 + 小驼峰(驼峰)命名
+                case SupportCase.DOT_CAMEL_CASE:    // 点分隔 + 小驼峰(驼峰)命名
                     if (isFirstWord) {
                         transformedWords.push(word);
                         if (isCurrentWordNormal) {
@@ -95,16 +102,19 @@ export function caseConversion(targetCase: SupportCase, str: string, eol: EOL, c
                 case SupportCase.SNAKE_PASCAL_CASE: // 下划线(蛇形) + 大驼峰(帕斯卡)命名
                 case SupportCase.KEBAB_PASCAL_CASE: // 中划线(连字符/脊柱式) + 大驼峰(帕斯卡)命名
                 case SupportCase.SPACE_PASCAL_CASE: // 空格分隔 + 大驼峰(帕斯卡)命名
+                case SupportCase.DOT_PASCAL_CASE:   // 点分隔 + 大驼峰(帕斯卡)命名
                     transformedWords.push(pascalCaseWord);
                     break;
                 case SupportCase.SNAKE_CASE:        // 下划线(蛇形)命名
                 case SupportCase.KEBAB_CASE:        // 中划线(连字符/脊柱式)命名
                 case SupportCase.SPACE_CASE:        // 空格分隔命名
+                case SupportCase.DOT_CASE:          // 点分隔命名x
                     transformedWords.push(word);
                     break;
                 case SupportCase.SNAKE_UPPER_CASE:  // 下划线(蛇形) + 全大写命名
                 case SupportCase.KEBAB_UPPER_CASE:  // 中划线(连字符/脊柱式) + 全大写命名
                 case SupportCase.SPACE_UPPER_CASE:  // 空格分隔 + 全大写命名
+                case SupportCase.DOT_UPPER_CASE:    // 点分隔 + 全大写命名
                     transformedWords.push(word.toUpperCase());
                     break;
                 default: